requirements.txt
faiss-cpu
sentence-transformers
transformers
PyPDF2
flask
pandas
numpy
pyyaml
